da vedere imu calib feedback

+++ SW_DP11_INIT/Core/Src/can.c
@@ -299,7 +299,7 @@ void CAN_receive(int ID, uint16_t firstInt, uint16_t secondInt, uint16_t thirdIn
 			  Indicators[HEAD_1].floatValore 	= ((int16_t)firstInt)/100.0;
 				Indicators[ACC_Z_1].floatValore = ((int16_t)secondInt)/100.0;
 				Indicators[GYR_Y_1].floatValore = ((int16_t)thirdInt)/10.0;
-				Indicators[IMU1_INFO].intValore = fourthInt;
+				Indicators[IMU1_INFO].intValore = (uint8_t)fourthInt;
 			 break;
 		 case	IMU2_DATA_1_ID:
 			 	Indicators[ACC_X_2].floatValore = ((int16_t)firstInt)/100.0;
@@ -311,7 +311,7 @@ void CAN_receive(int ID, uint16_t firstInt, uint16_t secondInt, uint16_t thirdIn
 			  Indicators[HEAD_2].floatValore  = ((int16_t)firstInt)/100.0;
 				Indicators[ACC_Z_2].floatValore = ((int16_t)secondInt)/100.0;
 				Indicators[GYR_Y_2].floatValore = ((int16_t)thirdInt)/10.0;
-				Indicators[IMU2_INFO].intValore = fourthInt;
+				Indicators[IMU2_INFO].intValore = (uint8_t)fourthInt;
 			 break;
 //		 case DAU_FR_ID:
 //			  Indicators[BPS_F].intValore  = thirdInt; //da togliere
